python_scripts/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
bitarray==3.7.1
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.2
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.7.0
hexbytes==1.3.1
idna==3.10
iniconfig==2.1.0
multidict==6.6.4
packaging==25.0
parsimonious==0.10.0
pluggy==1.6.0
poly_eip712_structs==0.0.1
propcache==0.3.2
py_clob_client==0.25.0
py_order_utils==0.3.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
pyunormalize==16.0.0
regex==2025.9.1
requests==2.32.5
rlp==4.1.0
tabulate==0.9.0
toolz==1.0.0
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
web3==7.13.0
websockets==15.0.1
yarl==1.20.1
